chore(lint): drop the redundant int64 conversion on NOT_APPLICABLE
spectypes.NOT_APPLICABLE is already declared int64, so the conversion is an
unconvert violation. It sits on main, so every branch inherits it and every
lint log carries a red error line — and because the step is continue-on-error,
that permanent error is what a contributor's own lint problem would hide behind.

+++ protocol/chainlib/chain_fetcher_addon_probe_test.go
@@ -95,7 +95,7 @@ func TestFetchLatestBlockNumKeepsBaseCollectionSemantics(t *testing.T) {
 
 	block, err := fetcher.FetchLatestBlockNum(context.Background())
 	require.Error(t, err)
-	require.Equal(t, int64(spectypes.NOT_APPLICABLE), block)
+	require.Equal(t, spectypes.NOT_APPLICABLE, block)
 }
 
 // acalaVerifications is the set GetVerifications returns for a node declaring
